ser error no  se arregla el problema
